requirements.txt
braceexpand==0.1.7
clip_benchmark==1.6.1
datasets==3.3.2
fsspec==2024.2.0
huggingface_hub==0.30.2
libauc==1.4.0
nltk==3.9.1
PyYAML==6.0.2
regex==2024.11.6
regex==2024.7.24
torch==2.4.0+cu124
torch==2.4.0
torchvision==0.19.0+cu124
tqdm==4.67.1
tqdm==4.66.5
transformers==4.44.2
webdataset==0.2.100
wilds==2.0.0
